refactor: Created a separate page for the home page
