+++ lib/main.dart
@@ -1,11 +1,8 @@
-import 'dart:developer';
-
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 
+import 'src/pages/home_page.dart';
 import 'src/supported_locale.dart';
-import 'src/widgets/locale_dropdown_button.dart';
-import 'src/widgets/recipe_input_form.dart';
 
 /// The title of this app.
 const appTitle = 'Ingredient Collector';
@@ -27,80 +24,20 @@ void main() async {
 }
 
 /// The [IngredientCollectorApp].
-class IngredientCollectorApp extends StatefulWidget {
+class IngredientCollectorApp extends StatelessWidget {
   /// Creates a new [IngredientCollectorApp].
   const IngredientCollectorApp({super.key});
 
   @override
-  State<IngredientCollectorApp> createState() => _IngredientCollectorAppState();
-}
-
-class _IngredientCollectorAppState extends State<IngredientCollectorApp> {
-  Future<void> _onLocaleChanged(Locale locale) async {
-    await context.setLocale(locale);
-    if (!mounted) return;
-    setState(() {
-      log("Locale changed to ${locale.languageCode}.");
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    var localeDropdownButton = LocaleDropdownButton(
-      // ignore: avoid-redundant-async, async is still necessary here
-      onChanged: (newValue) async => _onLocaleChanged(newValue.locale),
-    );
-
-    return MaterialApp(
-      home: Scaffold(
-        key: ValueKey(context.locale.languageCode),
-        appBar: AppBar(
-          title: const Text(appTitle),
-          actions: [localeDropdownButton],
+  Widget build(BuildContext context) => MaterialApp(
+        home: const HomePage(),
+        title: appTitle,
+        theme: ThemeData(
+          primarySwatch: Colors.blue,
         ),
-        body: SingleChildScrollView(
-          child: Container(
-            alignment: Alignment.topCenter,
-            child: const AppBody(),
-          ),
-        ),
-      ),
-      title: appTitle,
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      locale: context.locale,
-      localizationsDelegates: context.localizationDelegates,
-      supportedLocales: context.supportedLocales,
-      debugShowCheckedModeBanner: false,
-    );
-  }
-}
-
-/// The body of this app.
-class AppBody extends StatelessWidget {
-  /// Creates a new [AppBody].
-  const AppBody({super.key});
-
-  double _getContainerWidth(double width) {
-    if (width >= 1400) return 1320;
-    if (width >= 1200) return 1140;
-    if (width >= 992) return 960;
-    if (width >= 768) return 720;
-    if (width >= 576) return 540;
-    return width - 20;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    var queryData = MediaQuery.of(context);
-    var deviceWidth = queryData.size.width;
-    var containerWidth = _getContainerWidth(deviceWidth);
-
-    return Container(
-      width: containerWidth,
-      margin: const EdgeInsets.symmetric(vertical: 30),
-      child: const RecipeInputForm(),
-    );
-  }
+        locale: context.locale,
+        localizationsDelegates: context.localizationDelegates,
+        supportedLocales: context.supportedLocales,
+        debugShowCheckedModeBanner: false,
+      );
 }
